% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFilePath.R
\name{myFilePath}
\alias{myFilePath}
\title{A function to convert file path text on clipboard to standard R format}
\usage{
myFilePath()
}
\description{
This converts \ to \\ from text on clipboard
}
\examples{
myFilePath()
}
\keyword{file}
\keyword{path}
