% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLibrary.R
\name{myLibrary}
\alias{myLibrary}
\title{A function to install packages if they do not exist}
\usage{
myLibrary(myPackages)
}
\arguments{
\item{myPackages}{List all packages needed. There is no default.}
}
\description{
This function checks if needed packages are installed and installs and loads
all requested packages. Additionally, the package warnings are suppressed.
}
\examples{
myPackages <- c('dplyr','ggplot2')
myLibrary(myPackages)
}
\keyword{install.packages}
\keyword{library,}
