% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAutoWD.R
\name{setAutoWD}
\alias{setAutoWD}
\title{This function sets the working directory to the source file location.
Must be using RStudio}
\usage{
setAutoWD(...)
}
\description{
This function sets the working directory to the source file location.
Must be using RStudio
}
\examples{
setAutoWD
}
\keyword{directory,}
\keyword{set}
\keyword{setwd}
\keyword{working}
