% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLibrary.R
\name{myLibrary}
\alias{myLibrary}
\title{A function to install packages if they do not exist}
\usage{
myLibrary(...)
}
\description{
This function checks if needed packages are installed and installs and loads
all requested packages.
}
\examples{
myLibrary('dplyr','ggplot2')
}
\keyword{directory,}
\keyword{set}
\keyword{setwd}
\keyword{working}
