% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLibrary.R
\name{myLibrary}
\alias{myLibrary}
\title{A function to install packages if they do not exist.}
\usage{
myLibrary(my.packages)
}
\arguments{
\item{my.packages}{List all packages needed. There is no default.}
}
\description{
This function checks if needed packages are installed and installs and loads all requested packages.
}
\examples{
my.packages <- c('dplyr','ggplot2')
myLibrary(my.packages)
}
\keyword{install.packages}
\keyword{library,}
